Add chapter 2 exercises
Chapter 2 show cases variables, different types of variable like
vars which are basic variable, vals which are variable that cannot
be reassigned twice (kinda like a constant) and actual constants with
the 'const val' modifier. It also the different primitives types of
the language and hwo to inspect the generated byte code if needed.