Add SimVillage class for learning anonymous function
I've learnt what is anonymous function (or a lambda) and how they
be used as arguments or how they be returned by other function.
This chapter also talked about the implicit return for lambdas,
function inline to have better performance, Type inference support
which like passing hash to the function so it improves readability
and the 'it' syntax when there's only on argument for the function.